Removed Attendee notes field and added attendeehistory.
Now Add Note functionality will append an entry in the attendeehistory table for that user, instead of just being a text field that anyone can edit.

If the notes field exists in the data import, it will be inserted in to the attendee history table with the current date and attributed to the user that imports it.

The notes field on Order is still text (and will likely remain that way)