Removed unnecessary error logging in BadgePresenter
BadgePresenter would log an error when someone tried to view a null badge. The problem is, that would happen normally when you closed the badge being viewed. The table's selection gets cleared, which triggers the event listener with a null badge entry.

Just stopped logging an error for that. Fixes #30